requirements.txt
django-reversion==1.7
django-treebeard==1.61
pil==1.1.7